+++ metadata.rb
@@ -4,11 +4,10 @@ license          'All rights reserved'
 name             'cybera_sensu_server'
 description      'Installs/Configures a sensu server for Cybera'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.2.1' # Sprint 2015/03
+version          '0.4.0' # Sprint 2015/05
 
 recipe "default",         ""
 
 depends "sensu"
 depends "uchiwa"
-depends "monitoring"
 depends "nginx"