+++ CHANGES.md
@@ -1,1 +1,3 @@
-
+# cybera_sensu_server 0.4.0 (2015-05 sprint)
+* removed dependency on the moodle monitoring cookbook
+* this no longer sets up the sensu-client (leave that to a different cookbook)